test(backend): allow routes tests to run without DB
